// src/models/inquiries/Inquiries.ts
export interface InquiriesModel {
  userSenderId: number;
  userReceiverId: number;
  propertyId: number;
  firstname: string;
  lastname: string;
  address: string;
  streetAddress: string;
  addressLine2: string;
  stateProvince: string;
  postalZipCode: string;
  phoneNumber: string;
  email: string;
  inquiryDescription: string;
}